+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Using original argument names in mapper in order to avoid possible problems with labeled arguments [#59](https://github.com/arturogutierrez/Openticator/pull/59)
 * AES Encrypting support using SpongyCastle [#60](https://github.com/arturogutierrez/Openticator/pull/60)
 * Stateless interactors [#62](https://github.com/arturogutierrez/Openticator/pull/62)
+* Removing abstract layout resource from BaseActivity [#63](https://github.com/arturogutierrez/Openticator/pull/63)
 
 ### 0.9.3 (13/12/2016)